+++ .gitignore
@@ -7,8 +7,9 @@ __pycache__/
 venv/
 env/
 
-
-
 # GC
 .venv/
 .env
+
+*.xlsx
+*.ipynb
