+++ backend/conf/funciones.py
