--- backend/fly.toml
@@ -1,22 +1,0 @@
-# fly.toml app configuration file generated for demo-beneficios-back on 2024-08-05T19:04:57-04:00
-#
-# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
-#
-
-app = 'demo-beneficios-back'
-primary_region = 'scl'
-
-[build]
-
-[http_service]
-  internal_port = 8001
-  force_https = true
-  auto_stop_machines = 'stop'
-  auto_start_machines = true
-  min_machines_running = 0
-  processes = ['app']
-
-[[vm]]
-  memory = '1gb'
-  cpu_kind = 'shared'
-  cpus = 1
