--- frontend/ejemplo.py
@@ -1,2 +1,0 @@
-print("Hola todos")
-print("adios")
