--- frontend/fly.toml
@@ -1,22 +1,0 @@
-# fly.toml app configuration file generated for demo-beneficios on 2024-08-05T18:58:21-04:00
-#
-# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
-#
-
-app = 'demo-beneficios'
-primary_region = 'scl'
-
-[build]
-
-[http_service]
-  internal_port = 8502
-  force_https = true
-  auto_stop_machines = 'stop'
-  auto_start_machines = true
-  min_machines_running = 0
-  processes = ['app']
-
-[[vm]]
-  memory = '1gb'
-  cpu_kind = 'shared'
-  cpus = 1
